Improving the Sorting Functionality
