+++ Repository/Extensions/RepositoryEmployeeExtensions.cs
@@ -4,6 +4,7 @@ using System.Linq.Dynamic.Core;
 using System.Reflection;
 using System.Text;
 using Entities.Models;
+using Repository.Extensions.Utility;
 
 namespace Repository.Extensions;
 
@@ -23,36 +24,12 @@ public static class RepositoryEmployeeExtensions
     }
     public static IQueryable<Employee> Sort(this IQueryable<Employee> employees, string orderByQueryString)
     {
-        if(string.IsNullOrEmpty(orderByQueryString))
+        if(string.IsNullOrWhiteSpace(orderByQueryString))
         {
             return employees.OrderBy(entity =>  entity.Name);
         }
 
-        var orderParams = orderByQueryString.Trim().Split(',');
-
-        var propInfos = typeof(Employee).GetProperties(BindingFlags.Public | BindingFlags.Instance);
-
-        var orderQueryBuilder = new StringBuilder();
-
-        foreach( var param in orderParams)
-        {
-            if(string.IsNullOrWhiteSpace(param))
-            {
-                continue;
-            }
-
-            var propFromQueryName = param.Split(" ")[0];
-            var objProp = propInfos.FirstOrDefault(propInfo => propInfo.Name.Equals(propFromQueryName, StringComparison.OrdinalIgnoreCase));
-            if(objProp is null)
-            {
-                continue;
-            }
-
-            var direction = param.EndsWith(" desc") ? "descending" : "ascending";
-            orderQueryBuilder.Append($"{objProp.Name.ToString()} {direction},");
-        }
-        
-        var orderQuery = orderQueryBuilder.ToString().TrimEnd(',',' ');
+        var orderQuery = OrderQueryBuilder.CreateOrderQuery<Employee>(orderByQueryString);
 
         if(string.IsNullOrWhiteSpace(orderQuery))
         {
