+++ Repository/Extensions/Utility/OrderQueryBuilder.cs
@@ -1,0 +1,42 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+using Entities.Models;
+
+namespace Repository.Extensions.Utility;
+
+public static class OrderQueryBuilder
+{
+    public static string CreateOrderQuery<T>(string orderByQueryString)
+    {
+        var orderParams = orderByQueryString.Trim().Split(',');
+
+        var propInfos = typeof(T).GetProperties(BindingFlags.Public | BindingFlags.Instance);
+
+        var orderQueryBuilder = new StringBuilder();
+
+        foreach (var param in orderParams)
+        {
+            if (string.IsNullOrWhiteSpace(param))
+            {
+                continue;
+            }
+
+            var propFromQueryName = param.Split(" ")[0];
+            var objProp = propInfos.FirstOrDefault(propInfo => propInfo.Name.Equals(propFromQueryName, StringComparison.InvariantCultureIgnoreCase));
+            if (objProp is null)
+            {
+                continue;
+            }
+
+            var direction = param.EndsWith(" desc") ? "descending" : "ascending";
+            orderQueryBuilder.Append($"{objProp.Name.ToString()} {direction},");
+        }
+
+        var orderQuery = orderQueryBuilder.ToString().TrimEnd(',', ' ');
+        return orderQuery;
+    }    
+}
